% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_things_sim_data.R
\name{all_things_sim_data}
\alias{all_things_sim_data}
\title{all_things_sim_data -- samples, censors, stipw's -- gets outsample_sim() up and running quickly}
\usage{
all_things_sim_data(sim_seed, sample_size, sim_slope, sim_intercept,
  sim_ses_coef, sim_age_coef)
}
\arguments{
\item{sim_seed}{passed to set.seed()}

\item{sample_size}{(defaults to 1000) and must be a multiple of 100}

\item{sim_slope}{see slope in calculate_ses()}

\item{sim_intercept}{see intercept in calculate_ses()}

\item{sim_ses_coef}{see ses_coef in apply_censoring()}

\item{sim_age_coef}{see age_coef in apply_censoring()}
}
\value{
return(list(a =   wtd_trajectories, b = unwtd_fncs, c =   wtd_fncs, d = observed_with_stipw, e =      all_with_stipw, f = censored, g = age_vec))
}
\description{
not implemented in insample_sim() (yet).  Same inputs as in/outsample_sim()
}
\examples{
---
}

