% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{calculate_ses}
\alias{calculate_ses}
\title{Sample the prepped data of the boys of the Berkeley Growth Study Data}
\usage{
calculate_ses(data_in = NULL)
}
\arguments{
\item{data_in}{an object returned from sample_data()}
}
\description{
Using the oversampling of "boys", the returned object from "sample_data()";
imputes a made-up linear relationship between fabricated SES variable (think household income)
and the final height in column 32 of data_in.  In future I will build in
params for the 100 and the 12 and the rnorm in teh SES assignment.
}
\examples{
d<-prep_data()
head(d)
over_samp_mat<-sample_data(d,1000)
long_ses <- calculate_ses(over_samp_mat)
}
\keyword{oversampling}

