% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/calculate_stipw.R
\name{calculate_stipw}
\alias{calculate_stipw}
\title{Calculate STIPW based on propensity for being study}
\usage{
calculate_stipw(data_in = NULL, na.action = "omit")
}
\arguments{
\item{data_in}{an object returned from apply_censoring()}

\item{na.action}{"keep" means the dataset returned will be same number of rows as data_in and "omit" discards all data past censoring observation}
}
\value{
a dataset similar to data_in with stipw in columns; and potentially fewer rows or NA-filled rows
for induced censoring.
}
\description{
Names are hard coded, so stick
to script in the examples.
}
\examples{
\dontrun{
d<-prep_data()
head(d)
over_samp_mat<-sample_data(d,1000)
with_ses <- calculate_ses(over_samp_mat)
long <-make_long(with_ses)
head(long)
censored <- apply_censoring(long)
head(censored,18)
observed_with_stipw <- calculate_stipw(censored,"keep")
}
}

