% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/calculate_stipw_generic.R
\name{calculate_stipw_generic}
\alias{calculate_stipw_generic}
\title{Calculate STIPW based on propensity for being study}
\usage{
calculate_stipw_generic(data_in = NULL, na.action = "omit")
}
\arguments{
\item{data_in}{a data.table (see the data.table package, Tutorial, and FAQs) that is in long format of
at least the id, time, and outcome and some covariates for the probability weights.}

\item{na.action}{"keep" means the dataset returned will be same number of rows as data_in and "omit" discards all data past censoring observation}
}
\value{
a data.table similar to data_in with stipw in columns; and potentially fewer rows or NA-filled rows
for induced censoring.
}
\description{
Whereas in calculate_stipw() the names are hardcoded for the Berkeley Study simulation;
We aim here to generalize the stipw for the HIV application.  Thus the name
calculate_stipw_generic().
}
\examples{
## revisit this example/flow for HIV application...
d<-prep_data()
head(d)
over_samp_mat<-sample_data(d,1000)
with_ses <- calculate_ses(over_samp_mat)
long <-make_long(with_ses)
head(long)
censored <- apply_censoring(long)
head(censored,18)
observed_with_stipw <- calculate_stipw(censored,"keep")
}

