% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/calculate_wtd_trajectories.R
\name{calculate_wtd_trajectories}
\alias{calculate_wtd_trajectories}
\title{Calculate the ST-STIPW based on for graphical viewing as well as input to fpca methods}
\usage{
calculate_wtd_trajectories(data_in = NULL)
}
\arguments{
\item{data_in}{an object returned from calculate_stipw()}
}
\value{
a dataset similar to data_in with stipw in columns; and potentially fewer rows or NA-filled rows
for induced censoring.
}
\description{
Takes a calculate_stipw(,na.action="omit") object and calculates the weighted trajectories
where the weights are standardized (within observation time) STIPWs so that
the (regular, unweighted) mean of the weighted trajectories is the same as the
weighted mean of the original trajectories.
Names are hard coded, so stick
to script in the examples.
}
\examples{
d<-prep_data()
head(d)
over_samp_mat<-sample_data(d,1000)
with_ses <- calculate_ses(over_samp_mat)
long <-make_long(with_ses)
head(long)
censored <- apply_censoring(long)
head(censored,18)
observed_with_stipw <- calculate_stipw(censored,"omit")
wtd_trajectories <- calculate_wtd_trajectories(observed_with_stipw)
head(wtd_trajectories)
}

