% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insample_sim_hh2.R
\name{insample_sim_hh2}
\alias{insample_sim_hh2}
\title{Hopkins Hybrid 1: Run a simulation and then perform prediction.  Based on insample_sim().  Censoring is
fabricated a la the SES of the Growth data.  We will make a _hh2 to do a more realistic censoring.}
\usage{
insample_sim_hh2(sim_seed = 101, sample_size = 1000, sim_slope = 100,
  sim_intercept = 12, sim_ses_coef = 0.01, sim_age_coef = 0.01,
  censoring_coef = 1, hh_rds = "./data_local/hopkins_hybrid.RDS",
  hh_rds_simplog = "./data_local/hopkins_hybrid_prob_ltfu_coeffs.RDS",
  hh_rds_long = "./data_local/hopkins_hybrid_long.RDS")
}
\arguments{
\item{sim_seed}{passed to set.seed()}

\item{sample_size}{(defaults to 1000) and must be a multiple of 100}

\item{sim_slope}{see slope in calculate_ses()}

\item{sim_intercept}{see intercept in calculate_ses()}

\item{sim_ses_coef}{see ses_coef in apply_censoring()}

\item{sim_age_coef}{see age_coef in apply_censoring()}

\item{censoring_coef}{used to multiply the prob.cens from the simplog in hh2 runs. Default value is 1.  will try 0.95 and 1.05 for range.}

\item{hh_rds}{the path in single quotes to the local copy of hopkins_hybrid.RDS}
}
\value{
results a data frame with rmse for each approach for that simulated dataset
}
\description{
This function comes after many months of running readme_sim and talks with Bryan Lau on
what we need to demonstrate for the method.
}
\examples{
---
}

