% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/make_long.R
\name{make_long}
\alias{make_long}
\title{Make up SES (household income) relationship with final height for oversampled data}
\usage{
make_long(data_in = NULL)
}
\arguments{
\item{data_in}{an object returned from calculate_ses()}
}
\description{
Takes wide format with individual covariates and makes it long while
redoing the ids (original, oversampled ids retained as 'id', 'new.id' is unique) for a sensical melt.  Names are hard coded, so stick
to script in the examples.
}
\examples{
d<-prep_data()
head(d)
over_samp_mat<-sample_data(d,1000)
with_ses <- calculate_ses(over_samp_mat)
long <-make_long(with_ses)
head(long)
}
\keyword{oversampling}

