% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/outsample_sim.R
\name{outsample_sim}
\alias{outsample_sim}
\title{!!!!Achtung: Weighted methods are not done and are omitted.  BJS 5/20/2015  -- Run a simulation and then perform prediction.  Based on insample_sim() / insample_sim.R}
\usage{
outsample_sim(sim_seed = 101, sample_size = 1000, sample_size2 = 1,
  sim_slope = 100, sim_intercept = 12, sim_ses_coef = 0.01,
  sim_age_coef = 0.01)
}
\arguments{
\item{sim_seed}{passed to set.seed()}

\item{sample_size}{(defaults to 1000) and must be a multiple of 100}

\item{sample_size2}{(defaults to 1) the out of sample sample size}

\item{sim_slope}{see slope in calculate_ses()}

\item{sim_intercept}{see intercept in calculate_ses()}

\item{sim_ses_coef}{see ses_coef in apply_censoring()}

\item{sim_age_coef}{see age_coef in apply_censoring()}
}
\value{
results a data frame with rmse for each approach for that simulated dataset
}
\description{
This function comes after many months of running readme_sim and talks with Bryan Lau on
what we need to demonstrate for the method.
}
\examples{
---
}

