% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prep the Berkeley Growth Study Data}
\usage{
prep_data()
}
\arguments{
\item{NONE}{}
}
\description{
This returns "boys".
}
\examples{
d<-prep_data()
head(d)
}
\keyword{NONE}

