% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/readme_sim.R
\name{readme_sim}
\alias{readme_sim}
\title{Run a simulation as displayed in the readMe}
\usage{
readme_sim(sim_seed = 101, sample_size = 1000, sim_slope = 100,
  sim_intercept = 12, sim_ses_coef = 0.1, sim_age_coef = 0.2)
}
\arguments{
\item{sim_seed}{passed to set.seed()}

\item{sample_size}{(defaults to 1000) and must be a multiple of 100}

\item{sim_slope}{see slope in calculate_ses()}

\item{sim_intercept}{see intercept in calculate_ses()}

\item{sim_ses_coef}{see ses_coef in apply_censoring()}

\item{sim_age_coef}{see age_coef in apply_censoring()}
}
\value{
results a data frame with rmse for each approach for that simulated dataset
}
\description{
Runs all the commands in the readMe to generate rmse for each method.
Run several instances of this to do a "simulation study"
}
\examples{
---
}

