% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{sample_data}
\alias{sample_data}
\title{Sample the prepped data of the boys of the Berkeley Growth Study Data}
\usage{
sample_data(data_in = boys, tot_subj = 1000)
}
\arguments{
\item{data_in}{an object returned from prep_data()}

\item{tot_subj}{total number of subjects sampled with replacement -- numeric, multiple of 100.}
}
\description{
This returns an oversampling of "boys", the returned object from "prep_data()".
}
\examples{
d<-prep_data()
head(d)
overSampMat<-sample_data(d,1000)
}
\keyword{oversampling}

