% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/sample_data_fpc.R
\name{sample_data_fpc}
\alias{sample_data_fpc}
\title{Sample the prepped data of the boys of the Berkeley Growth Study Data
based on the fpc as opposed to the oversampling in "sample_data()"}
\usage{
sample_data_fpc(data_in = NULL, tot_subj = 1000, seed = 101,
  timepoints = age_vec)
}
\arguments{
\item{data_in}{an object returned from prep_data()}

\item{tot_subj}{total number of subjects sampled with replacement -- numeric, multiple of 100.}

\item{seed}{an integer for set.seed(seed) to assist reproducible simulations}

\item{timepoints}{the grid that the columns of data_in represent}
}
\description{
This returns an oversampling of "boys", the returned object from "prep_data()".
}
\examples{
d<-prep_data()
head(d)
overSampMat<-sample_data(d,1000)
}
\keyword{oversampling}

