% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{apply_censoring}
\alias{apply_censoring}
\title{Apply censoring scheme that is related to SES}
\usage{
apply_censoring(data_in = NULL)
}
\arguments{
\item{data_in}{an object returned from calculate_ses()}
}
\description{
Takes output from make_long() and applies censoring
by creating a made-up, linear, hard-coded relationship between the probability
of being censored at each observation time and SES
Names are hard coded, so stick
to script in the examples.
}
\examples{
d<-prep_data()
head(d)
over_samp_mat<-sample_data(d,1000)
with_ses <- calculate_ses(over_samp_mat)
long <-make_long(with_ses)
head(long)
censored <- apply_censoring(long)
head(censored)
}

