% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{calculate_ses}
\alias{calculate_ses}
\title{Make up SES (household income) relationship with final height for oversampled data}
\usage{
calculate_ses(data_in = NULL, slope = 100, intercept = 12)
}
\arguments{
\item{data_in}{an object returned from sample_data()}

\item{slope}{a number that determines the linear relationship of final height and SES.
For instance, if slope=100 and intercept=12, then the SES for individual i is
SES(i) = 100*( height(i) - minimum of all heights )/(max of all heights - min of all heights) + 12 + rnorm(1)}

\item{intercept}{a number, see explanation of slope}
}
\description{
Using the oversampling of "boys", the returned object from "sample_data()";
imputes a made-up linear relationship between fabricated SES variable (think household income)
and the final height in column 32 of data_in.  In future I will build in
params for the 100 and the 12 and the rnorm in teh SES assignment.
}
\examples{
d<-prep_data()
head(d)
over_samp_mat<-sample_data(d,1000)
with_ses <- calculate_ses(over_samp_mat)
}
\keyword{oversampling}

