% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{prep_data}
\alias{prep_data}
\title{Prep the Berkely Growth Study Data}
\usage{
prep_data(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

